factor out subevent.py
